Move local calendar text fixtures to conftest.py
